added some additional helpe methods to driver.h/driver.cc for internal driver communications moved fiducial drivers to new message API - now compile, but untested moved camerav4l to new API, also compiles but is untested

+++ server/drivers/camera/v4l/camerav4l.cc
@@ -168,7 +168,7 @@ class CameraV4L : public Driver
   private: virtual void Main();
 
   // Process incoming messages from clients 
-  int ProcessMessage(ClientData * client, player_msghdr * hdr, uint8_t * data, uint8_t * resp_data, int * resp_len);
+  int ProcessMessage(ClientData * client, player_msghdr * hdr, uint8_t * data, uint8_t * resp_data, size_t * resp_len);
 
   // Process requests.  Returns 1 if the configuration has changed.
   //private: int HandleRequests();
@@ -418,7 +418,7 @@ void CameraV4L::Main()
 
 ////////////////////////////////////////////////////////////////////////////////
 // Process an incoming message
-int CameraV4L::ProcessMessage(ClientData * client, player_msghdr * hdr, uint8_t * data, uint8_t * resp_data, int * resp_len)
+int CameraV4L::ProcessMessage(ClientData * client, player_msghdr * hdr, uint8_t * data, uint8_t * resp_data, size_t * resp_len)
 {
   assert(hdr);
   assert(data);
